refactor(bloomctl): group cyl commands by data type (bloomctl cyl download)
Move the top-level `download` command under the `cyl` group so cylinder
commands are grouped by data type, matching the legacy CLI (`bloom cyl
download`) and the existing `bloomctl cyl ingest-result`.

Reorganize the source one file per command, mirroring legacy's cyl/ tree:
- src/bloomctl/cyl/__init__.py  — the `cyl` group + command registration
- src/bloomctl/cyl/download.py  — download command + its logic (was download.py)
- src/bloomctl/cyl/ingest.py    — ingest-result command + its logic (was ingest.py)
cli.py now holds only the root group, `login`, and `_authed_client`.

Update test imports (bloomctl.download/ingest -> bloomctl.cyl.*), the
`cyl download` invocation in the download tests, README, and CHANGELOG.
Add a help-tree assertion that `download` lives under `cyl`, not top level.
