+++ bloomcli/src/bloomctl/cyl/download.py
@@ -1,4 +1,4 @@
-"""Cylinder experiment download: metadata (scans.csv) + per-frame images.
+"""`bloomctl cyl download`: metadata (scans.csv) + per-frame images.
 
 Pure helpers (column mapping, paths) are separated from the supabase/storage I/O
 so the contract is unit-testable without a live server.
@@ -10,6 +10,10 @@ import csv
 from dataclasses import dataclass
 from pathlib import Path
 from typing import Any
+
+import click
+
+from ..credentials import DEFAULT_PROFILE
 
 # scans.csv schema: (output column, source key in a cyl_scans_extended row).
 # Order matches the legacy CLI's predict-container contract; `genotype` is
@@ -209,3 +213,140 @@ def write_download_log(result: DownloadResult, path: Path) -> None:
         lines.append(line)
     lines.append(f"\nSummary: {result.ok} downloaded, {result.failed} failed, {result.total} total")
     path.write_text("\n".join(lines) + "\n", encoding="utf-8")
+
+
+# --- command ----------------------------------------------------------------
+
+
+@click.command(name="download")
+@click.argument("out_dir", type=click.Path(file_okay=False, path_type=Path))
+@click.option(
+    "--experiment-id",
+    "--experiment_id",
+    "experiment_id",
+    type=int,
+    default=None,
+    help="Download a whole experiment by ID (mutually exclusive with --scan-id).",
+)
+@click.option(
+    "--scan-id",
+    "--scan_id",
+    "scan_id",
+    type=int,
+    default=None,
+    help="Download a single scan by ID (mutually exclusive with --experiment-id).",
+)
+@click.option(
+    "-p",
+    "--profile",
+    default=DEFAULT_PROFILE,
+    show_default=True,
+    help="Credentials profile to use.",
+)
+@click.option(
+    "--meta-only",
+    "--meta_only",
+    "meta_only",
+    is_flag=True,
+    help="Write scans.csv only; skip image download.",
+)
+@click.option(
+    "--plant-qr-code",
+    "--plant_qr_code",
+    "plant_qr_code",
+    default=None,
+    help="Restrict to a single plant QR code.",
+)
+@click.option(
+    "--plant-age-min",
+    "--plant_age_min",
+    "plant_age_min",
+    type=int,
+    default=0,
+    show_default=True,
+    help="Minimum plant age in days.",
+)
+@click.option(
+    "--plant-age-max",
+    "--plant_age_max",
+    "plant_age_max",
+    type=int,
+    default=1000,
+    show_default=True,
+    help="Maximum plant age in days.",
+)
+@click.option(
+    "--limit",
+    type=int,
+    default=100000,
+    show_default=True,
+    help="Maximum number of scans to fetch.",
+)
+def download(
+    out_dir: Path,
+    experiment_id: int | None,
+    scan_id: int | None,
+    profile: str,
+    meta_only: bool,
+    plant_qr_code: str | None,
+    plant_age_min: int,
+    plant_age_max: int,
+    limit: int,
+) -> None:
+    """Download a cylinder experiment (--experiment-id) or a single scan (--scan-id):
+    metadata (scans.csv) and per-frame images."""
+    from .. import auth
+    from ..credentials import load_credentials
+
+    # Exactly one of --experiment-id / --scan-id.
+    if (experiment_id is None) == (scan_id is None):
+        raise click.UsageError("Pass exactly one of --experiment-id or --scan-id.")
+
+    # NOTE: this inlines what `_authed_client(profile)` does; kept inline here to
+    # avoid overlapping PR #385, which migrates `download` to the shared helper.
+    try:
+        creds = load_credentials(profile)
+    except (FileNotFoundError, ValueError) as exc:
+        raise click.ClickException(f"{exc} — run `bloomctl login`.") from exc
+    try:
+        client = auth.make_authed_client(creds)
+    except auth.AuthError as exc:
+        raise click.ClickException(str(exc)) from exc
+
+    if scan_id is not None:
+        scan = fetch_scan(client, scan_id)
+        if scan is None:
+            raise click.ClickException(f"Scan {scan_id} not found.")
+        scans = [scan]
+    else:
+        scans = fetch_scans(
+            client,
+            experiment_id,
+            plant_qr_code=plant_qr_code,
+            plant_age_min=plant_age_min,
+            plant_age_max=plant_age_max,
+            limit=limit,
+        )
+    genotypes = fetch_genotypes(client, [s.get("accession_id") for s in scans])
+    rows = [build_scan_row(s, genotypes.get(s.get("accession_id"))) for s in scans]
+
+    out = Path(out_dir)
+    csv_path = out / "scans.csv"
+    write_scans_csv(rows, csv_path)
+    click.echo(f"Wrote {len(rows)} scans -> {csv_path}")
+
+    if meta_only:
+        return
+
+    result = download_images(client, scans, out)
+    log_path = out / "download_log.txt"
+    write_download_log(result, log_path)
+    click.echo(
+        f"Downloaded {result.ok}/{result.total} image frames -> {out / 'images'}  (log: {log_path})"
+    )
+    if result.failed:
+        # Partial download: surface it and exit non-zero so a pipeline knows the
+        # output is incomplete (the log lists every failed frame).
+        raise click.ClickException(
+            f"{result.failed} of {result.total} frames failed to download — see {log_path}"
+        )
